Removing tox becasue tox will not be useful for simplestack.
The main reason for that is because tox uses sdist from setup.py to
generate the project structure and it doesn't include some files
(etc/simplestack.cfg and usr/sbin/simplestack) that are important to
build the project.

It also doesn't solve the libvirt module problem to use a virtualenv
environment, so we'll keep using nose directly.